Final Submission

+++ src/ukf.h
@@ -14,7 +14,7 @@ using Eigen::VectorXd;
 class UKF {
 public:
 
-  float previous_timestamp_;
+  long previous_timestamp_;
   ///* initially set to false, set to true in first call of ProcessMeasurement
   bool is_initialized_;
